pyproject.toml
[build-system]
requires = ['setuptools']
build-backend = 'setuptools.build_meta'

[project]
name = 'aclman'
version = '0.3.0'
description = "ACL manager middleware for IDeATe at Carnegie Mellon University"
readme = "README.md"
authors = [
  { name = 'Tim Parenti', email = 'tparenti@andrew.cmu.edu' }
]
requires-python = '>=3.8'
dependencies = ['pyyaml']

[project.urls]
repository = "https://github.com/CMUIDeATe/aclman"

[tool.setuptools]
# Use automatic discovery

[tool.setuptools.package-data]
data = ['*.csv']
